Implemented GameStudyDropdown
